LPS-34567 Merge changes of pages without conflicts in their friendly urls.
The existence of conflicts between the friendly urls of some pages in the site and the site template will not prevent the rest of the pages to be merged. The UUIDs of the site pages with conflicts in their URLs will be stored so the site admin can access this information and decide whether to change their urls so they can be merged again or not.