LPS-37546 Update default submit function (_defaultSubmitFormFn) to support "image" type input elements
The input element with a type attribute whose value is "image" represents either an image from which the UA enables a user to interactively select a pair of coordinates and submit the form, or alternatively a button from which the user can submit the form.

Based on this definition, I do not see any cons to do not recognise <input type="image"... /> elements on submit.

Details: http://www.w3.org/TR/html-markup/input.image.html